fix: restrict quantity to be positive number

+++ orders/models.py
@@ -34,7 +34,7 @@ class Cart(BaseModel):
 class CartProduct(BaseModel):
     carts = models.ForeignKey(to=Cart, on_delete=models.CASCADE, related_name='cartproducts')
     products = models.ForeignKey(to=Product, related_name='carts', on_delete=models.CASCADE)
-    quantity = models.IntegerField()
+    quantity = models.PositiveSmallIntegerField()
     
     class Meta:
         unique_together = [['carts', 'products']]
